Update TODO list
Immutable board class won't be needed. There is no need to keep that info.
The turn number will suffice. It would be needed should the board be drawn
more than one time.

Guess checker is done.

About reading parameters in a more civilised way... As long as there is no
real use of this code it's difficult to guess which way would be the best.
Probably delegating them to the constructor would be a good idea but I don't
think it should be done now. Should I write GUI, I will do something with that.